LayoutTests/accessibility/alt-tag-on-image-with-nonimage-role-expected.txt


This tests the alternative text calculation when setting a role on an img with an alt attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS text.stringValue is 'AXValue: TEST1'
PASS platformValueForW3CName(group) is "TEST2"
PASS platformValueForW3CName(button) is "TEST3"
PASS successfullyParsed is true

TEST COMPLETE
